+++ libs/iso-bmff/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './CONTAINERS.ts'
+export * from './fourCcToUint32.ts'
 export * from './isContainer.ts'
 export * from './isFullBox.ts'